+++ packages/xchain-litecoin/CHANGELOG.md
@@ -1,3 +1,13 @@
+# v.x.x.x
+
+### Update
+
+- Uses Bitaps to submit transactions instead of Sochain
+
+### Fix
+
+- Fix `getExplorerUrl` to bitaps.
+
 # v.0.1.0 (2021-02-19)
 
 ### Breaking change